requirements.txt
absl-py==1.4.0
appnope==0.1.3
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.54.0
h5py==3.8.0
idna==3.4
install==1.3.5
ipykernel==6.22.0
ipython==8.13.1
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
JsonForm==0.0.2
jsonschema==4.18.1
jsonschema-specifications==2023.6.1
JsonSir==0.0.2
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.12.0
keras-tuner==1.3.5
kiwisolver==1.4.4
kt-legacy==1.0.5
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mpmath==1.3.0
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==4.22.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.5.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
Python-EasyConfig==0.1.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
referencing==0.29.1
requests==2.30.0
requests-oauthlib==1.3.1
Resource==0.2.1
rpds-py==0.8.10
rsa==4.9
scikit-learn==1.2.2
scikit-optimize @ git+https://github.com/scikit-optimize/scikit-optimize.git@a2369ddbc332d16d8ff173b12404b03fea472492
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-addons==0.20.0
tensorflow-docs @ git+https://github.com/tensorflow/docs@9ecb223ba8b830ee1c336fc8378af2282cc8be17
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-macos==2.12.0
tensorflow-metal==1.0.0
termcolor==2.3.0
threadpoolctl==3.1.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.1
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==2.3.3
wrapt==1.14.1
